Android: Step comments (#1104)
^ALTAPPS-1283

+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/core/view/ui/widget/compose/ShimmerModifier.kt
@@ -95,7 +95,7 @@ private fun Modifier.infiniteShimmer(
                 animation = tween(
                     durationMillis = durationMillis,
                     easing = easing,
-                    delayMillis = delayMillis.toInt()
+                    delayMillis = delayMillis
                 ),
                 repeatMode = RepeatMode.Restart,
             ),
